Add support for retrieving fund logs by school class, enhance DTOs with parent-related data, and refactor fund log queries for improved flexibility.

+++ src/main/java/com/example/schoolmoney/domain/fundlog/FundLogView.java
@@ -12,7 +12,7 @@ public interface FundLogView {
 
     String getChild_full_name();
 
-    Double getAmount_in_cents();
+    Long getAmount_in_cents();
 
     String getCurrency();
